migrate active section and card side effect to deck page level
