+++ client/src/features/deck/routes/Deck.tsx
@@ -1,15 +1,33 @@
 import React, { useState, useEffect } from 'react'
+import { useHistory, useLocation, useParams } from 'react-router-dom'
 import Page from '../../../ui/Page'
 import Card from '../Card'
 import DeckNav from '../DeckNav'
+import { useDeck } from '../deckSlice'
 import NavSettings from '../NavSettings'
 
 function Deck(props: any) {
   const [loading, setLoading] = useState(true)
+  const history = useHistory()
+  const location = useLocation()
+  const params: any = useParams()
+  const {
+    activeSection,
+    activeCard,
+  } = useDeck()
 
   useEffect(() => {
     setLoading(false)
   }, [])
+
+  useEffect(() => {
+    const back = activeCard.side === 'back' || location.search.includes('?back') ? '?back=true' : ''
+    if (activeSection.id && activeCard.id) {
+      history.push(
+        `/decks/${params.deckId}/${activeSection.id}/${activeCard.id}${back}`
+      )
+    }
+  }, [activeSection.id, activeCard.id])
 
   return (
     <Page loaded={!loading}>
